+++ regen/objects/model-transformation.cpp
