+++ regen/objects/model-transformation.h
