+++ regen/shapes/cull-shape.h
@@ -3,7 +3,7 @@
 
 #include <regen/scene/state.h>
 #include "regen/buffer/ssbo.h"
-#include "regen/states/model-transformation.h"
+#include "../objects/model-transformation.h"
 #include "spatial-index.h"
 
 namespace regen {
