+++ regen/simulation/boid-simulation.h
@@ -4,7 +4,7 @@
 #include "regen/shader/shader-input.h"
 #include "regen/shapes/bounds.h"
 #include "regen/textures/texture.h"
-#include "regen/states/model-transformation.h"
+#include "../objects/model-transformation.h"
 #include "regen/textures/collision-map.h"
 
 namespace regen {
