+++ regen/textures/material-state.cpp
