+++ regen/textures/screen-state.cpp
